// frontend/src/app/api/models/tour-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TourResponse {
  computedChildFriendlyScore?: (number | string);
  computedPopularityScore?: (number | string);
  description?: (string | null);
  distance?: (number | string);
  endLocation?: string;
  estimatedTime?: (number | string);
  id?: (number | string);
  mapImagePath?: (string | null);
  routeGeojson?: (string | null);
  startLocation?: string;
  title?: string;
  transportType?: string;
}